requirements.txt
Django>=4.2.0,<5.0.0
python-telegram-bot>=20.0,<21.0
gspread>=5.7.0,<6.0.0
google-auth>=2.16.0,<3.0.0
gunicorn>=20.1.0,<21.0.0